requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
cloudinary==1.29.0
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==3.3
mysqlclient==2.1.0
Pillow==9.1.0
psycopg2==2.9.3
pycodestyle==2.8.0
PyJWT==2.3.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0
gunicorn==20.1.0